+++ src/static/config.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2024, Salesforce, Inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+// eslint-disable-next-line import/prefer-default-export
+export const CUSTOM_API_DEFAULT_BASE_URI =
+  'https://{shortCode}.api.commercecloud.salesforce.com/custom/{apiName}/{apiVersion}';